Always create VMM domains
This is to help with setups where apic_provision_infra
is set to False because the user wants to set up
switch and entity profiles manually. However they have
no way of creating a VMM-domain in APIC. So we create
OpenStack VMM domain and let them associate the appropriate
entity profile to it manually.
Other VMM domains like VMware, Microsoft usually already
exist in APIC, so the 'create' for these cases simply
verifies the existence of these VMM domains.

Fixes #40

Signed-off-by: Amit Bose <amitbose@gmail.com>