Add config options to apicapi for physical nodes
* Added an option to indicate the encapsulation mode
to use for a VMM/physical domain ('encap_mode'). If
unspecified, the encap mode is guessed from other
config options like we do today.
* Added section 'apic_physical_network:<net-name>' which
can be used to associate individual server nodes to
logical segments in the physical network. The format
of options in this section is -
hosts=<comma-separated list of hosts>
segment_type=...

Partially-Closes noironetworks/support#113

Signed-off-by: Amit Bose <amitbose@gmail.com>